requirements.txt
Fabric==1.11.1
Flask==0.11
Flask-Login==0.3.2
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
SQLAlchemy==1.0.12
WTForms==2.1
Werkzeug==0.11.9
aniso8601==1.1.0
astroid==1.4.5
bcrypt==2.0.0
blinker==1.4
cffi==1.6.0
click==6.6
colorama==0.3.7
ecdsa==0.13
gunicorn==19.4.5
itsdangerous==0.24
lazy-object-proxy==1.2.2
paramiko==1.16.0
passlib==1.6.5
pycparser==2.14
pycrypto==2.6.1
pylint==1.5.5
python-dateutil==2.5.3
pytz==2016.4
requests==2.10.0
six==1.10.0
speaklater==1.3
wrapt==1.10.8
wsgiref==0.1.2
prettytable
salt-pepper==0.5
